# Make `flutter update-packages` run in parallel