## Description

This modifies the `flutter update-packages` and `flutter update-packages --force-upgrade` commands so that the 66 invocations of `dart pub get` in each repo project run in parallel instead of in series.

Recently, there was a regression of the speed of `pub get` (the Dart team is already working on a fix, so it's temporary), and that highlighted for me just how inefficient `update-packages` was for a command that we (Flutter repo devs) run daily. Our workstations have a lot of cores, and we could be using them. It seemed ripe for parallelism.

There was one snag, which is that if you have a cold cache (e.g. delete the `.pub-cache` directory), then if all of the `dart pub get` runs happen in parallel, you get contention when unpacking and moving any freshly downloaded packages into place in the cache, since there isn't any locking going on.

For this case, however, it is solvable: because all of our `pubspec.yaml` files have the same versions of all the packages: it now writes out a dummy file with the versions we currently depend on, and runs `dart pub get` on that before it starts trying to do version solving for all 66 packages in parallel. That way, the cache is warm by the time things happen in parallel, and there's no contention for unpacking.

For doing the concurrent tasks, I added a generic `TaskQueue` class that I originally wrote for Dartdoc.

The results are pretty nice: On my machine (72 cores), it goes from 4 minutes and 38 seconds to 30 seconds on a cold (empty) cache, and from 4 minutes and 27 seconds to 10 seconds on a warm cache. That's about a 30x speedup!

## Tests
 - Added tests for the `TaskQueue`.